Refactoring to use new Features system for core Server
You might think of it as changing one mess for another. While I feel that this is not perfect yet, I wanted to avoid the situation where each system has its own `Map<Client/GameType/something, SystemInternalData>`, so I feel right now that this is a step in the right direction but it's certainly not perfect.