Fix Coup Assassin cost 6 coins bug by changing order of rules
The bug happened because no player had an action to perform, which made rules be called twice and therefore another cost-rule added on top of the first. Changing the order of these rules checks for an empty stack *after* resolving an action and therefore makes sure that there is an action to be performed when the stack is empty.