Fix last(?) edge-case with copy fork #291 (see commit details for log)
From comments on GitHub issue #291: (copied here as GitHub may not live forever after all)

There is a slight conflict between "If you do this, you have to be able to do X later" and "Replay of moves"...

In order to look into the future, I copy the entire game and replays all moves. So I look into the past to look into the future, which will look into the past, which will look into the future, which will look into the past, which will look into the future, which will look into the past, which will look into the future...

This chain needs to be broken somehow, preferably by not looking into the future while looking into the past. Can the whole requires part be skipped, and preferably also the forking itself, since we know that it is okay to do these actions because they have already been made? Can we wrap the entire forkGame call in a forks section or something, that only gets invoked if game is not being replayed?

---

In some cases regarding "If you do this, you have to be able to do X later", copying the entire game is unnecessary and copying only the relevant parts of the model to evaluate that is enough.

However, as a generic case, it's good if copying the entire game is still supported, so that's what I am aiming for right now.

---

One bug I encountered was that some of the forked games were not stopped. That has now been fixed.

---

In my test scenario I added a game where you can change a simple int value, which starts at a random number from -20 to 20. In order to be allowed to increase by 10, the forked game - which adds an increase of 40 - must have a value of less than 50. So in practice, the value needs to be below 10 in order to be allowed to increase by 10.

One edge case I found is that a coroutine was kept running.
Which happened because the game had caused an exception to be thrown.
Which happened because the action definition DSL statement was added twice.
This happened because the list of action DSLs was not cleared in between, which it usually is.
Which happened because one action was not allowed to be performed.
Which happened because a clone didn't have the same starting randomness value as the original.
Which happened because the replay hadn't been fully performed.
Which happens when replays contain no actions.
So the root cause is that replays are not correctly performed when they are not containing actions.

---

And the root cause mentioned above happens because of this statement in ReplayingListener:

`if (moves >= replayMoveCount) return`

The edge-case is, of course, that moves == 0 and replayMoveCount == 0.

---

Now at least the tests are stable, however, a new issue has appeared: When a game is forked, it doesn't take the most recently made action into consideration. This is because the code to send the feedback for lastAction is in the start of GameFlowImpl.nextAction(), which is executed after the DSL which creates the fork.

One possibility is to make a ActionPerforming (or ActionAllowed?) step, which is created and sent as soon as possible.

However, a bad thing about that approach is that all the replayable state stored in between steps will not be copied, the reason why the feedback is sent in the beginning of nextAction() in the first place is because of #272

---

A simpler solution to the problem of not performing the last move is to send it the most recently performed move, which hasn't had its feedback sent yet but has been stored for future sending.

This will have the consequence of not fully replaying anything that is done after the fork happens, such as:

`game.value += replayable.int("stepLoop") { Random.Default.nextInt(-2, 2 + 1) }`

However, does it matter? I think it is okay, since it's impossible to copy a randomness that hasn't happened yet.

I was for a while thinking about if we should have a check to check if we're forked (The Good Place reference!! 😄), but I don't think that's needed.

---

The mess continues!

Randomness from start of game was moved to the beginning of a step call, as the fork happens before the call to nextAction.
However, when it is placed there the randomness from onStart is not part of the replay data.

It would help significantly if the randomness from onStart was treated in a more similar way to the randomness from performing actions.

---

Solution, naturally, was to treat the randomness from onStart in a more similar way to the randomness from performing actions.