Add support for events performing a default action when executed
When an event is executed it may perform some effect once the NORMAL priorities are executed. Thus, NORMAL and later priorities can be used for *after* an effect has taken place while EARLIEST to EARLY happens *before* an effect takes place.

The motivation for supporting this is to support all the following scenarios:
- "When you gain X, gain one more" (before)
- "Whenever you gain X, gain one Y" (after)
- "You cannot gain X" (before)